Refactor OptionsDB logger aspects into LoggerWOptionsDB.
This will allow components not at the app level, like test of UI,
universe components etc. to use the Logger without pulling in OptionsDB.
