+++ msvc2015/Common/Common.vcxproj
@@ -194,6 +194,7 @@
     <ClInclude Include="..\..\util\MultiplayerCommon.h" />
     <ClInclude Include="..\..\util\i18n.h" />
     <ClInclude Include="..\..\util\Logger.h" />
+    <ClInclude Include="..\..\util\LoggerWithOptionsDB.h" />
     <ClInclude Include="..\..\util\OptionsDB.h" />
     <ClInclude Include="..\..\util\OptionValidators.h" />
     <ClInclude Include="..\..\util\Order.h" />
@@ -255,6 +256,7 @@
     <ClCompile Include="..\..\util\MultiplayerCommon.cpp" />
     <ClCompile Include="..\..\util\i18n.cpp" />
     <ClCompile Include="..\..\util\Logger.cpp" />
+    <ClCompile Include="..\..\util\LoggerWithOptionsDB.cpp" />
     <ClCompile Include="..\..\util\OptionsDB.cpp" />
     <ClCompile Include="..\..\util\Order.cpp" />
     <ClCompile Include="..\..\util\OrderSet.cpp" />
