+++ util/Logger.cpp
@@ -1,7 +1,5 @@
 #include "Logger.h"
 
-#include "OptionsDB.h"
-#include "i18n.h"
 #include "Version.h"
 
 #include <boost/log/core.hpp>
@@ -48,60 +46,7 @@ namespace {
 
     // Compile time constant pointers to constant char arrays.
     constexpr const char* const log_level_names[] = {"debug", "info", "warn", "error"};
-
-    constexpr LogLevel default_sink_level = LogLevel::debug;
-    constexpr LogLevel default_source_level = LogLevel::info;
-
-    DiscreteValidator<std::string> LogLevelValidator() {
-        std::set<std::string> valid_levels = {"debug", "info", "warn", "error",
-                                              "DEBUG", "INFO", "WARN", "ERROR",
-                                              "0", "1", "2", "3" };
-        auto validator = DiscreteValidator<std::string>(valid_levels);
-        return validator;
-    }
-
-    constexpr auto exec_option_name_prefix = "logging.execs.";
-    constexpr auto source_option_name_prefix = "logging.sources.";
-
-    std::regex exec_name_regex("(?:logging\\.execs\\.)(\\S+)");
-    std::regex source_name_regex("(?:logging\\.sources\\.)(\\S+)");
-
-
-    /** Add the log threshold for a logger with \p full_option_name to OptionsDB and return the
-        threshold read from OptionsDB. */
-    LogLevel AddLoggerToOptionsDB(const std::string& full_option_name) {
-
-        // Determine the type of logger, executable default or channel.
-        std::smatch ematch;
-        std::regex_search(full_option_name, ematch, exec_name_regex);
-        bool is_an_exec_root_logger = !ematch.empty();
-
-        std::smatch cmatch;
-        std::regex_search(full_option_name, cmatch, source_name_regex);
-        bool is_a_channel_logger = !cmatch.empty();
-
-        if (!is_an_exec_root_logger && !is_a_channel_logger)
-            WarnLogger(log) << "Adding a logger to OptionsDB with an unknown prefix. " << full_option_name;
-
-        // Find the appropriate defaults.
-        auto default_level = to_string(default_source_level);
-        auto description = UserStringNop("OPTIONS_DB_LOGGER_SOURCE_LEVEL");
-
-        if (is_an_exec_root_logger) {
-            default_level = to_string(default_sink_level);
-            description =  UserStringNop("OPTIONS_DB_LOGGER_FILE_SINK_LEVEL");
-        }
-
-        // Create the new option if necessary.
-        if (!GetOptionsDB().OptionExists(full_option_name))
-            GetOptionsDB().Add<std::string>(
-                full_option_name, description, default_level, LogLevelValidator());
-
-        // Return the threshold from the db.
-        return to_LogLevel(GetOptionsDB().Get<std::string>(full_option_name));
-    }
 }
-
 
 
 std::string to_string(const LogLevel level)
@@ -205,14 +150,7 @@ void InitLoggingSystem(const std::string& logFile, const std::string& _default_e
     logging::core::get()->add_global_attribute("TimeStamp", attr::local_clock());
 
     // Initialize the internal logger
-    RegisterLoggerWithOptionsDB(FO_GLOBAL_LOGGER_NAME(log)::get(), "log");
-
-    // Setup the OptionsDB options for the file sink.
-    LogLevel options_db_log_threshold = AddLoggerToOptionsDB(
-        exec_option_name_prefix + f_default_exec_logger_name);
-
-    // Use the option to set the threshold of the default logger
-    SetLoggerThreshold("", options_db_log_threshold);
+    ConfigureLogger(FO_GLOBAL_LOGGER_NAME(log)::get(), "log");
 
     // Print setup message.
     auto date_time = std::time(nullptr);
@@ -220,7 +158,7 @@ void InitLoggingSystem(const std::string& logFile, const std::string& _default_e
     InfoLogger() << FreeOrionVersionString();
 }
 
-void RegisterLoggerWithOptionsDB(NamedThreadedLogger& logger, const std::string& name) {
+void ConfigureLogger(NamedThreadedLogger& logger, const std::string& name) {
     if (name.empty())
         return;
 
@@ -242,12 +180,6 @@ void RegisterLoggerWithOptionsDB(NamedThreadedLogger& logger, const std::string&
 
     logging::core::get()->add_sink(sink_frontend);
 
-    // Setup the OptionsDB options for this source.
-    LogLevel options_db_log_threshold = AddLoggerToOptionsDB(
-        source_option_name_prefix + name);
-
-    // Use the option.
-    SetLoggerThreshold(name, options_db_log_threshold);
     LoggerCreatedSignal(logger, name);
 
     InfoLogger(log) << "Added log source \"" << name << "\".";
@@ -265,74 +197,4 @@ void SetLoggerThreshold(const std::string& source, LogLevel threshold) {
     logging::core::get()->reset_filter();
     f_min_channel_severity[source] = threshold;
     logging::core::get()->set_filter(f_min_channel_severity);
-}
-
-
-void UpdateLoggerThresholdsFromOptionsDB() {
-    SetLoggerThresholds(LoggerExecutableOptionsLabelsAndLevels());
-    SetLoggerThresholds(LoggerSourceOptionsLabelsAndLevels());
-}
-
-namespace {
-
-    /** Returns the list of full option names, logger names and thresholds for loggers in
-        OptionsDB will \p prefix using \p prefix_regex.*/
-    std::set<std::tuple<std::string, std::string, LogLevel>> LoggerOptionsLabelsAndLevels(
-        const std::string prefix, const std::regex& prefix_regex) {
-        // Get a list of all of the potential loggers
-        std::set<std::string> loggers;
-        GetOptionsDB().FindOptions(loggers, prefix, true);
-
-        std::set<std::tuple<std::string, std::string, LogLevel>> retval;
-        for (const auto& full_option : loggers) {
-            // Find the option name
-            std::smatch match;
-            std::regex_search(full_option, match, prefix_regex);
-            if (match.empty()) {
-                ErrorLogger(log) << "Unable to find a logger name from option name \"" << full_option << "\"";
-                continue;
-            }
-            const auto& option_name = match[1];
-
-            // Add it to options db and get the option value
-            const auto option_value = AddLoggerToOptionsDB(full_option);
-
-            // Add to return value
-            retval.insert(std::make_tuple(full_option, option_name, option_value));
-
-            DebugLogger(log) << "Found " << full_option << ",  name = " << option_name << " value = " << option_value;
-        }
-
-        return retval;
-    }
-}
-
-/** Return the option names, labels and levels for all executables from OptionsDB. */
-std::set<std::tuple<std::string, std::string, LogLevel>> LoggerExecutableOptionsLabelsAndLevels()
-{ return LoggerOptionsLabelsAndLevels(exec_option_name_prefix, exec_name_regex); }
-
-/** Return the option names, labels and levels for all sources/channels from OptionsDB. */
-std::set<std::tuple<std::string, std::string, LogLevel>> LoggerSourceOptionsLabelsAndLevels()
-{ return LoggerOptionsLabelsAndLevels(source_option_name_prefix, source_name_regex); }
-
-/** Sets the logger thresholds from a list of options, labels and thresholds. */
-void SetLoggerThresholds(const std::set<std::tuple<std::string, std::string, LogLevel>>& fulloption_name_and_levels) {
-    for (const auto& fulloption_name_and_level : fulloption_name_and_levels) {
-        const auto& full_option = std::get<0>(fulloption_name_and_level);
-        const auto& name = std::get<1>(fulloption_name_and_level);
-        const auto& value = std::get<2>(fulloption_name_and_level);
-
-
-        // Update the option in OptionsDB if it already exists.
-        if (GetOptionsDB().OptionExists(full_option))
-            GetOptionsDB().Set(full_option, to_string(value));
-
-        // Set the logger threshold.
-
-        std::smatch match;
-        std::regex_search(full_option, match, exec_name_regex);
-        bool is_my_root_logger = (!match.empty() && match[1] == f_default_exec_logger_name);
-
-        SetLoggerThreshold((is_my_root_logger ? "" : name), value);
-    }
 }
